Add files via upload
This project, "Restaurant-Management-System", is a console-based application developed in Java that simulates the functionality of a restaurant. It includes the following key features:

User Authentication:
Allows users to sign up by creating an account with a username, password, and phone number.
Provides a login feature where users can authenticate using their username or phone number and password.

Menu System:
Offers separate Veg and Non-Veg menus with various dishes listed by ID and price.
Users can add dishes to their cart by selecting the corresponding dish ID.

Cart Functionality:
The selected dishes are added to a cart, and the total bill is calculated dynamically as items are added.

Bill Calculation and Checkout:
The application displays the items in the cart along with the total bill amount during checkout.

The project showcases basic OOP principles like the use of methods for specific operations (e.g., creating an account, adding to the cart), and it leverages collections such as ArrayList for managing the cart.